requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
aiosqlite==0.19.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Tapo / Kasa
python-kasa==0.7.3

# eWeLink (HTTP client)
httpx==0.26.0
aiohttp==3.9.3

# WebSocket
websockets==12.0

# Encryption for token storage
cryptography==42.0.2

# Frontend templating
jinja2==3.1.3

# Utilities
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1
structlog==24.1.0
